[Nexters/FALCON#96] coin contract -> hammer contract, hammer 지급 및 갯수조회 기능 추가
